Add alias command to sort the output of the DU command
The `du` Unix command used to estimate the file space usage in the
current disk with combination with the parameters `-s` and `-h` prints
a human-readable summary of the space occupied by the files in the
current directory, however, this summary is unsorted and it is hard to
determine which files and/or folders are taking more space than others
during a manual cleanup. This commit adds a custom global command that
sorts and formats the output.