Fixed extract alias to prevent volatile globbling
The alias command used to simplify the extraction of archives in
different formats was assumming that the file name was composed of a
sequence of alphanumeric characters without spaces, this clearly means
that if you pass a file name with special characters and / or white
spaces the command would fail. This commit wraps the parameter in
between double quotes to prevent globbing.