Added more snippets for the PHPUnit autocompletion
PHPUnit methods like assertNan, assertArraySubset, and assertInfinite
were added with this commit after the update of the PHP installation and
the overall stack. The first method fails if the variable is not a NAN
which is refered as non-a-number; the second method fails if the input
variable is not a subset of the expected data, this is, if the expected
data is an array with three keys and the input is an array with three or
less of these keys the test will suceed, otherwise it will fail. The
third method fails if the variable is not an INF.