Added new PHPUnit methods for the deprecated exceptions
This adds the method expectException that can be used to set the
expected PHP exception name, expectExceptionCode can be used to set the
expected exception code / number, expectExceptionMessage can be used to
set the expected message, and expectExceptionMessageRegExp can be used
to set the expected message using a regular expression. Notice that all
these new methods can be used as individual tags in the Docblock for
each test case; the functions that were used in PHP < 5.3 are deprecated
but still available until the next version bump.