Added alias to colorize man command using less pager
Termcap is a database of terminal capabilities, so there needs to be a
database in which you can look up whichever terminal the user has in
order to find out what that kind of terminal is capable of and how to
tell it to do each thing. Each capability has a two-letter identifier,
and maps to a string of characters:

mb - Start blinking
md - Start bold mode
me - End all mode like so, us, mb, md and mr
so - Start standout mode
se - End standout mode
us - Start underlining
ue - End underlining

So when we want something underlined, for example, the "us" and "ue"
entries in our terminal's termcap record are what we need to send to the
terminal to start and end underlining that section of text. Since the
"man" command uses "less" to page the output of a document before it is
rendered we can colorize the output using this technique.

Details at https://news.ycombinator.com/item?id=12296000