Added binary path for the protobuf binary files
Protocol Buffers is a method of serializing structured data. It is
useful in developing programs to communicate with each other over a wire
or for storing data. The method involves an interface description
language that describes the structure of some data and a program that
generates source code from that description for generating or parsing a
stream of bytes that represents the structured data. Google developed
Protocol Buffers for use internally and has provided a code generator
for multiple languages under an open source license.