Add binary path for IPFS a P2P hypermedia protocol
We believe the internet has become humanity's most important technology.
We build protocols, systems, and tools to improve how it works. Today,
we are focused on distributed information systems. There are massive
inefficiencies in how we store, locate, and move information. Recent
advances have shown how to do things better. We are deploying new
systems based on them.

We are making the distributed web. Our systems improve how the web
works, based on protocols like Git, DHTs, BitTorrent, Bitcoin, and SFS.
We are focused on providing (a) a data-locality aware, secure network
stack for distributed applications, and (b) a model for ambient and
network-efficient storage.

[1] https://ipfs.io/
[2] http://ipn.io/