Modify command to print the extended Xterm palette
Some coding standard changes were applied to this command line tool to
respond to warnings triggered by the shellcheck Bash linter. The
conditionals were simplified to exit immediately after the execution of
their internal code and some changes to the exit code numbers were added
to report the correct status at the end of the execution of the tool.