Add mongodb to the list of installed packages
MongoDB (from humongous) is a free and open-source cross-platform
document-oriented database program. Classified as a NoSQL database
program, MongoDB uses JSON-like documents with schemas. MongoDB is
developed by MongoDB Inc. and is free and open-source, published under a
combination of the GNU Affero General Public License and the Apache
License.

- https://www.mongodb.com/