Add BIND to the list of installed packages
BIND is the most widely used Domain Name System (DNS) software on the
Internet. On Unix-like operating systems it is the de facto
standard.[citation needed] The software was originally designed at the
University of California Berkeley (UCB) in the early 1980s. The name
originates as an acronym of Berkeley Internet Name Domain, reflecting
the application's use within UCB. The software consists, most
prominently, of the DNS server component, called named, a contracted
form of name daemon. In addition the suite contains various
administration tools, and a DNS resolver interface library.

- https://www.isc.org/downloads/bind/