Add id3v2 to the list of installed packages
ID3 is a metadata container most often used in conjunction with the MP3
audio file format. It allows information such as the title, artist,
album, track number, and other information about the file to be stored
in the file itself.

There are two unrelated versions of ID3: ID3v1 and ID3v2. ID3v2 is
structurally very different from ID3v1, consisting of an extensible set
of "frames" located at the start of the file, each with a frame
identifier (a three- or four-byte string) and one piece of data.

ID3 is a de facto standard for metadata in MP3 files; no standardization
body was involved in its creation nor has such an organization given it
a formal approval status.[1] It competes with the APE tag in this arena.

- http://id3v2.sourceforge.net/
- https://en.wikipedia.org/wiki/ID3