Add command to download and import music to iTunes
Download a video from the Internet using youtube-dl and extracts the
audio before the file is imported to the local iTunes library. Some
metadata is assigned to the file before the import action, things like
the author's name and song title will be attached to the file via ID3v2
tags. Notice that this script only supports macOS.