Fix command to set the modification date of an image
On macOS there is a different tool that allows you to change the
creation time of a file, this information is also used by iPhotos to
order the images and videos imported into the library. Setting a
modification date and/or time while the creation time is older will
force iPhotos to import the image or video into the time specified in
the second factor.