Modify custom rules for the PSR2 command with exceptions
Minified files are going to be ignored and the custom ruleset file will
be created respecting the new lines. Now the XML code will be properly
formatted. Additionally, the PHPCS tool will now work recursively rather
than locally, this is to cover all the PHP files inside the current
directory.